Issue #377 - use the VMUSB drivers in place rather than as copies
Import the CAEN V977 support from the vmusb/daqconfig rather than
rely on an unmaintainable copy.
