+++ main/mvlc/Makefile.am
@@ -75,11 +75,12 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMTDC32.h $(IMPORT_DIR)/CMTDC32.cpp \
 	$(IMPORT_DIR)/CNADC2530.h $(IMPORT_DIR)/CNADC2530.cpp \
 	$(IMPORT_DIR)/CCAENV1x90Opcodes.h $(IMPORT_DIR)/CCAENV1x90Registers.h \
-	$(IMPORT_DIR)/CV1x90.cpp $(IMPORT_DIR)/CV1x90.h
+	$(IMPORT_DIR)/CV1x90.cpp $(IMPORT_DIR)/CV1x90.h \
+	$(IMPORT_DIR)/ CV977.h $(IMPORT_DIR)/CV977.cpp 
+
 
 
 LOCAL_MODULES= 	MADC32Registers.h \
-	 CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
 	CXLM.h CXLM.cpp CXLMTimestamp.cpp CXLMTimestamp.h \
 	CXLMFERA.cpp CXLMFERA.h \
